LogClassifier/requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
groq==0.9.0
sentence-transformers==3.3.1
joblib==1.3.2
pandas==2.0.2
scikit-learn==1.6.0
gradio==4.26.0
uvicorn>=0.22.0
pydantic>=2.0
requests>=2.31.0